+++ cspell.config.cjs
@@ -1,3 +1,4 @@
+// eslint-disable-next-line import/extensions
 const { createConfig } = require("./packages/cspell-config/index.js");
 
 module.exports = createConfig();