+++ packages/cspell-config/cspell.config.js
@@ -1,3 +1,4 @@
+// eslint-disable-next-line import/extensions
 const { createConfig } = require("./index.js");
 
 module.exports = createConfig();