+++ packages/eslint-config/eslint.config.js
@@ -1,3 +1,4 @@
+// eslint-disable-next-line import/extensions
 import config from "./index.js";
 
 export default config;