+++ packages/eslint-plugin-boundaries-e2e/test/fixtures/basic/src/modules/user/unknown-import.js
@@ -1,0 +1,6 @@
+// This should be INVALID - module cannot import unknown elements
+import unknown from "../../unknown/unknown-element/unknown.js";
+
+export const userModule = () => {
+  return unknown;
+};