+++ packages/eslint-plugin-boundaries/CHANGELOG.md
@@ -13,6 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## Unreleased
 
+### Added
+
+- feat: Add full TypeScript typings for all plugin settings and rules. Export main `Config` type and individual rule/setting option types. Support custom plugin name prefixes.
+- feat: Export constants and type guard methods for settings keys, rule names, and other configuration-related values.
+- feat: Add new createConfig helper that enforces valid types, automatically adds the plugin, and includes JS/TS file patterns. Support custom plugin name prefixes.
+- feat: Add export path `/config`, exporting createConfig helper and constants/types.
+- feat: Validate nodes. Print a warning in case a dependency node defined selected with the `additional-dependency-nodes` setting is not valid Literal.
+- feat: Add internal validations. Display message with repo issues link when an unexpected situation is found.
+
 ## [5.0.2] - 2025-10-06
 
 ### Changed